Fix CI build

+++ src/genie/read/lowlatency/decoder.cc
@@ -76,8 +76,8 @@ core::record::Chunk Decoder::decode_common(core::AccessUnit&& t) {
                 }
                 qvCounter++;
             }
-            if(!r.getSegments().empty()) {
-                r.setQVDepth(r.getSegments().front().getQualities().size());
+            if (!r.getSegments().empty()) {
+                r.setQVDepth(static_cast<uint8_t>(r.getSegments().front().getQualities().size()));
             }
         }
     }
